backend_api/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-storage
firebase-admin
httpx[http2]